+++ src/scouty.rs
@@ -37,27 +37,48 @@ use async_std::task;
 use log::{error, info, warn};
 use std::{convert::TryInto, result::Result, str::FromStr, thread, time};
 use subxt::{
-    ext::sp_core::crypto, storage::StorageKey, utils::AccountId32, OnlineClient,
-    PolkadotConfig,
+    backend::{
+        legacy::{rpc_methods::StorageKey, LegacyRpcMethods},
+        rpc::RpcClient,
+    },
+    ext::sp_core::crypto,
+    utils::AccountId32,
+    OnlineClient, PolkadotConfig,
 };
 
-pub async fn create_substrate_node_client(
+pub async fn _create_substrate_node_client(
     config: Config,
 ) -> Result<OnlineClient<PolkadotConfig>, subxt::Error> {
     OnlineClient::<PolkadotConfig>::from_url(config.substrate_ws_url).await
 }
 
+pub async fn create_substrate_rpc_client_from_config(
+    config: Config,
+) -> Result<RpcClient, subxt::Error> {
+    RpcClient::from_url(config.substrate_ws_url).await
+}
+
+pub async fn create_substrate_client_from_rpc_client(
+    rpc_client: RpcClient,
+) -> Result<OnlineClient<PolkadotConfig>, subxt::Error> {
+    OnlineClient::<PolkadotConfig>::from_rpc_client(rpc_client).await
+}
+
 pub async fn create_or_await_substrate_node_client(
     config: Config,
-) -> (OnlineClient<PolkadotConfig>, SupportedRuntime) {
+) -> (
+    OnlineClient<PolkadotConfig>,
+    LegacyRpcMethods<PolkadotConfig>,
+    SupportedRuntime,
+) {
     loop {
-        match create_substrate_node_client(config.clone()).await {
-            Ok(client) => {
-                let chain = client.rpc().system_chain().await.unwrap_or_default();
-                let name = client.rpc().system_name().await.unwrap_or_default();
-                let version = client.rpc().system_version().await.unwrap_or_default();
-                let properties =
-                    client.rpc().system_properties().await.unwrap_or_default();
+        match create_substrate_rpc_client_from_config(config.clone()).await {
+            Ok(rpc_client) => {
+                let rpc = LegacyRpcMethods::<PolkadotConfig>::new(rpc_client.clone());
+                let chain = rpc.system_chain().await.unwrap_or_default();
+                let name = rpc.system_name().await.unwrap_or_default();
+                let version = rpc.system_version().await.unwrap_or_default();
+                let properties = rpc.system_properties().await.unwrap_or_default();
 
                 // Display SS58 addresses based on the connected chain
                 let chain_prefix: ChainPrefix =
@@ -87,7 +108,19 @@ pub async fn create_or_await_substrate_node_client(
                     chain, config.substrate_ws_url, name, version
                 );
 
-                break (client, SupportedRuntime::from(chain_token_symbol));
+                match create_substrate_client_from_rpc_client(rpc_client.clone()).await {
+                    Ok(client) => {
+                        break (client, rpc, SupportedRuntime::from(chain_token_symbol));
+                    }
+                    Err(e) => {
+                        error!("{}", e);
+                        info!(
+                            "Awaiting for connection using {}",
+                            config.substrate_ws_url
+                        );
+                        thread::sleep(time::Duration::from_secs(6));
+                    }
+                }
             }
             Err(e) => {
                 error!("{}", e);
@@ -101,12 +134,13 @@ pub async fn create_or_await_substrate_node_client(
 pub struct Scouty {
     runtime: SupportedRuntime,
     client: OnlineClient<PolkadotConfig>,
+    rpc: LegacyRpcMethods<PolkadotConfig>,
     matrix: Matrix,
 }
 
 impl Scouty {
     async fn new() -> Scouty {
-        let (client, runtime) =
+        let (client, rpc, runtime) =
             create_or_await_substrate_node_client(CONFIG.clone()).await;
 
         // Initialize matrix client
@@ -119,12 +153,17 @@ impl Scouty {
         Scouty {
             runtime,
             client,
+            rpc,
             matrix,
         }
     }
 
     pub fn client(&self) -> &OnlineClient<PolkadotConfig> {
         &self.client
+    }
+
+    pub fn rpc(&self) -> &LegacyRpcMethods<PolkadotConfig> {
+        &self.rpc
     }
 
     /// Returns the matrix configuration
@@ -214,7 +253,7 @@ fn spawn_and_restart_subscription_on_error() {
 }
 
 pub fn get_account_id_from_storage_key(key: StorageKey) -> AccountId32 {
-    let s = &key.0[key.0.len() - 32..];
+    let s = &key[key.len() - 32..];
     let v: [u8; 32] = s.try_into().expect("slice with incorrect length");
     v.into()
 }